+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 - Fix TargetProcess url formatting in settings
 - Fix ToolWindow naming
 - Fix ToolWindow reloading
-- 
+
 ## [0.2.0] - 2024-02-07
 
 ### Fixed